Added PancakePuzzle. Improved a lot of the error-checking for the new SearchEnvironment methods. Split up the TurnTaking simulation code into its own piece and cleaned up a lot of this code. Made the GenericStepAlgorithms use the new SearchEnvironment methods.

+++ trunk/apps/dynamicsearch/GenericStepAlgorithm.h
@@ -46,23 +46,23 @@ public:
 	virtual int GetPath(environment *env, state from, state to, std::vector<action> &path) = 0;
 	virtual const char *GetName() = 0;
 
-	virtual unsigned long long GetNodesExpanded() = 0; // number of nodes expanded
-	virtual unsigned long long GetNodesTouched() = 0; // number of nodes generated
+	virtual uint64_t GetNodesExpanded() = 0; // number of nodes expanded
+	virtual uint64_t GetNodesTouched() = 0; // number of nodes generated
 
 	/**
 	Returns the number of nodes on which a goal test has been performed. Note, in many cases (like A*)
 	this value will be equivalent to the number of nodes expanded.
 	**/
-	virtual unsigned long long GetNodesChecked() = 0;
+	virtual uint64_t GetNodesChecked() = 0;
 
 	/**
 	Sets limits for the various metrics of efficiency. Enter 0 for no limit.
 	Returns true if setting the limit succeeds,
 	and fails otherwise.
 	**/
-	virtual bool SetExpandedLimit(unsigned long long limit) = 0;
-	virtual bool SetTouchedLimit(unsigned long long limit) = 0;
-	virtual bool SetCheckedLimit(unsigned long long limit) = 0;
+	virtual bool SetExpandedLimit(uint64_t limit) = 0;
+	virtual bool SetTouchedLimit(uint64_t limit) = 0;
+	virtual bool SetCheckedLimit(uint64_t limit) = 0;
 
 	virtual void LogFinalStats(StatCollection *stats) = 0;
